ocpn_frame: Move S57Load and related stuff out of it.
It should not be necessary to include ocpn_frame.h in order to
use S57Load(). Move things out of ocnp_frame.
