+++ gui/include/gui/s57_load.h
@@ -1,0 +1,24 @@
+/***************************************************************************
+ *   Copyright (C) 2010 - 2025 by David S. Register                        *
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License as published by  *
+ *   the Free Software Foundation; either version 2 of the License, or     *
+ *   (at your option) any later version.                                   *
+ *                                                                         *
+ *   This program is distributed in the hope that it will be useful,       *
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+ *   GNU General Public License for more details.                          *
+ *                                                                         *
+ *   You should have received a copy of the GNU General Public License     *
+ *   along with this program; if not, see <https://www.gnu.org/licenses/>. *
+ **************************************************************************/
+
+#ifndef S_57_LoaD_h
+#define S_57_LoaD_h
+
+/** Load the s57 stuff and make global ps52plib ready to be used. */
+void LoadS57();
+
+#endif  // S_57_LoaD_h
