+++ gui/src/ocpn_plugin_gui.cpp
@@ -73,11 +73,11 @@
 #include "toolbar.h"
 #include "waypointman_gui.h"
 
-extern PlugInManager* s_ppim;  // FIXME (leamas) another name for global mgr
-
 #if wxUSE_XLOCALE || !wxCHECK_VERSION(3, 0, 0)
 extern wxLocale* plocale_def_lang;
 #endif
+
+extern PlugInManager* s_ppim;  // FIXME (leamas) another name for global mgr
 
 extern options* g_pOptions;  // FIXME (leamas) merge to g_options
